feature: [#7] migrate project to typescript
