+++ package.json
@@ -20,8 +20,9 @@
 	},
 	"devDependencies": {
 		"@eslint/js": "^9.21.0",
-		"@types/react": "^19.0.10",
-		"@types/react-dom": "^19.0.4",
+		"@types/node": "^22.15.21",
+		"@types/react": "^19.1.5",
+		"@types/react-dom": "^19.1.5",
 		"@vitejs/plugin-react": "^4.3.4",
 		"dotenv-webpack": "^8.1.0",
 		"eslint": "^9.21.0",
@@ -29,6 +30,7 @@
 		"eslint-plugin-react-refresh": "^0.4.19",
 		"globals": "^15.15.0",
 		"prettier": "^3.5.3",
-		"vite": "^6.2.5"
+		"typescript": "^5.8.3",
+		"vite": "^6.3.5"
 	}
 }
