--- src/components/Button/Button.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import './Button.css';
-
-export function Button({ children, onClick }) {
-	return (
-		<button className="custom-button" onClick={onClick}>
-			{children}
-		</button>
-	);
-}
