--- src/data/footerConfObj.js
@@ -1,14 +1,0 @@
-export const footerSections = [
-	{
-		title: 'COMPANY',
-		links: ['Home', 'Order', 'FAQ', 'Contact'],
-	},
-	{
-		title: 'TEMPLATE',
-		links: ['Style Guide', 'Changelog', 'Licence', 'Webflow University'],
-	},
-	{
-		title: 'FLOWBASE',
-		links: ['More Cloneables'],
-	},
-];
