+++ src/data/footerConfObj.ts
@@ -1,0 +1,19 @@
+interface FooterSection {
+  title: string;
+  links: string[];
+}
+
+export const footerSections: FooterSection[] = [
+  {
+    title: 'COMPANY',
+    links: ['Home', 'Order', 'FAQ', 'Contact'],
+  },
+  {
+    title: 'TEMPLATE',
+    links: ['Style Guide', 'Changelog', 'Licence', 'Webflow University'],
+  },
+  {
+    title: 'FLOWBASE',
+    links: ['More Cloneables'],
+  },
+];
