--- src/data/options.js
@@ -1,6 +1,0 @@
-export const options = [
-	{ id: '', label: 'Home' },
-	{ id: 'menu', label: 'Menu' },
-	{ id: 'company', label: 'Company' },
-	{ id: 'login', label: 'Login' },
-];
