+++ src/data/options.ts
@@ -1,0 +1,11 @@
+interface NavOption {
+  id: string;
+  label: string;
+}
+
+export const options: NavOption[] = [
+  { id: '', label: 'Home' },
+  { id: 'menu', label: 'Menu' },
+  { id: 'company', label: 'Company' },
+  { id: 'login', label: 'Login' },
+];
