+++ src/firebase/firebaseApp.ts
@@ -3,4 +3,4 @@ import { getAuth } from 'firebase/auth';
 import { firebaseConfig } from './firebaseConfig';
 
 const app = initializeApp(firebaseConfig);
-export const auth = getAuth(app);
+export const auth = getAuth(app);
