--- src/pages/CartPage/CartPage.jsx
@@ -1,5 +1,0 @@
-import './CartPage.css';
-
-export function CartPage() {
-	return <p>this is your Cart</p>;
-}
