+++ src/pages/CartPage/CartPage.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import './CartPage.css';
+
+export function CartPage(): React.ReactElement {
+  return <p>this is your Cart</p>;
+}
