--- src/pages/CompanyPage/CompanyPage.jsx
@@ -1,5 +1,0 @@
-import './CompanyPage.css';
-
-export function CompanyPage() {
-	return <p>this is your Company</p>;
-}
