+++ src/pages/CompanyPage/CompanyPage.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import './CompanyPage.css';
+
+export function CompanyPage(): React.ReactElement {
+  return <p>this is your Company</p>;
+}
